backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
email-validator==2.1.0
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2
requests==2.31.0
groq==0.4.1
google-generativeai==0.3.2